requirements.txt
faiss-cpu
numpy
ollama
